Added Email confirmation, forgot password, and lockout functionality
I added an email confirmation that is sent using SendGrid that is
required to click prior to a new user gaining access to the application.
The forgot password functionality has also been added where it will send
an email to the user that is making the request.   Lockout functionality
was enabled so that after 3 incorrect attempts the user account is
locked out and a message is then displayed to the user informing them of
their locked out status.