Added rough draft google map to the home tab. Added forced authorization on all tabs
Added a rough draft of the google maps api on to the home page.  This
allows for a destination and a source to be plotted and will return the
miles and distance as well as the ideal route.  User login is now
prompted upon entering the site and is required in order to access any
of the tabs.